feat: add separate manager and employee dashboards
