+++ app/employee/page.tsx
@@ -1,5 +1,5 @@
-import Dashboard from '@/components/dashboard';
+import EmployeeDashboard from '@/components/employee-dashboard';
 
 export default function EmployeePage() {
-  return <Dashboard initialTab="employee" singleView />;
+  return <EmployeeDashboard />;
 }
